kubecon-us-18-iotcore/requirements.txt
paramiko
PyYAML
requests
delegator.py
nuclio_sdk
google-api-python-client==1.7.4
google-auth-httplib2==0.0.3
google-auth==1.5.1
google-cloud-pubsub==0.38.0
